create ttl util provider
